requirements.txt
accelerate==1.2.1
bitsandbytes==0.45.0
fastapi==0.115.6
langchain==0.3.14
langchain-google-genai==2.0.8
matplotlib==3.10.0
numpy==2.2.1
pandas==2.2.3
PyMuPDF==1.25.1
sentence-transformers==3.3.1
spacy==3.8.3
torch==2.5.1+cu124
tqdm==4.67.1
transformers==4.48.0
uvicorn==0.34.0
langchain-google-genai
langchain-huggingface

pydantic~=2.10.5
langchain-community~=0.3.14
langchain-text-splitters~=0.3.5
python-dotenv~=1.0.1
chromadb~=0.6.3